Moved reused components to a folder within the saved-jobs directory
