+++ src/app/saved-jobs/components/ButtonWithBadge.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../../components/ui/button'
+import { Button } from '../../../components/ui/button'
 
 interface ButtonWithBadgeProps {
     badgeText: string | number;
